Database :: Fill CsharpEdges table
This commit fixes the RelationCollector class
and makes changes to the Program.cs.
We add a ConcurrentDictionary in order to
gather data into it with the relationcollector.
After data is collected, the contents of this
ConcurrentDictionary is transferred to the DbContext
and the changes are saved.
RelationCollector needs further work on the
VisitValueDeclarator and VisitInvocationExpression
functions.