Diagrams :: Data conversion from string to vector<FileId>
The data we gather from the Csharp component in order
to build the diagram needs to be converted in order to
be useful. This means converting from a string which contains
both the uses and the revuses to two vector<FileId> instances
to later pass on to the diagram generator functions.